Refactor potentially shared code out of cli script
